refactor(chat): extract llm_client.py (provider adapters + config resolution)
Verbatim move of the provider URL/model tables, Anthropic wire-format
adapters (_is_anthropic/_build_anthropic_*/_convert_messages_for_anthropic),
reasoning-model token headroom, _byok_error_message, _detect_model_from_url,
and the _resolve_llm_config / _resolve_with_model_override /
_resolve_fallback_llm_config / _platform_provider_matches chain.

_build_llm_http_client and the ensure_public_https_url_resolves DNS-guard
call deliberately STAY in chat.py (tests patch app.services.chat.httpx and
app.services.chat.ensure_public_https_url_resolves for the pinned-transport
path). Moved symbols are re-imported into app.services.chat so api-layer
and eval imports keep working.

Two test files updated to follow the moved collaborators: settings and
decrypt_api_key are now read in llm_client, so test_llm_catalog and the
two byok resolve tests patch llm_client.{settings,decrypt_api_key}
instead of chat.* (the DNS-guard patch stays on chat, its call site did).

chat.py 1900 -> 1056 LOC; five focused modules extracted. 819 tests pass.
